requirements.txt
flask==2.0.3
werkzeug==2.0.3
openai==0.27.0
python-dotenv==1.0.0